must-gather-ai-analysis/document_training_system/requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
PyPDF2>=3.0.0
python-docx>=0.8.11
pyyaml>=6.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
huggingface-hub>=0.17.0
tokenizers>=0.14.0
scikit-learn>=1.3.0